refactor: unification du composant BadgeTransmissionDonnees
